Some basic stuff
Haven't really try to check the logic.  Still stuck with generating
strategies for an agent. All 3 strategies have the same pattern. Need to
design more reliable random number generator. Or another solution would
be create a fix number of strategies and assign an agents to the
strategy, 3 strategies each agent.